Ai-Based PowerGrid/backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.0
mysql-connector-python==8.3.0
PyJWT==2.8.0
redis==5.0.3
rq==1.16.1
